# docs/api.openapi.yaml
openapi: 3.1.0
info:
  title: SubnetIQ REST and Streaming API
  version: 1.0.0
  description:
    'Guest mathematics and lookups; Supabase bearer-token persistence; typed assistant SSE. Requests
    accept application/json and are limited to 512 KiB, except project POST/PATCH envelopes (2.25 MiB) with
    a 2 MiB plan cap. Success bodies are returned directly. Failure bodies follow Error. API responses default
    to Cache-Control: no-store. Global quota is 180 requests/minute per IP identity. Production quotas are
    atomic PostgreSQL RPC calls. No credentials means honest guest mode in development; protected operations
    return 503. All ordinary user database access retains the verified user RLS context.'
servers:
  - url: http://localhost:3001
    description: Local API
  - url: /
    description: Current deployment
...
paths:
...
  /api/ai/status:
    get:
      operationId: get_ai_status
      summary: Read actual AI adapter configuration
      tags:
        - Assistant
      security: []
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
        '400':
          description: Invalid request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '429':
          description: Quota exceeded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          headers:
            Retry-After:
              schema:
                type: integer
              description: Seconds until retry.
        '503':
          description: Required credentials, migrations, or service are unavailable.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      description:
        Reports provider, requestedProvider, configured, mode, authenticationRequired, persistenceConfigured,
        dailyLimit, message, and explicit adapter availability.
  /api/ai/chat:
    post:
      operationId: post_ai_chat
      summary: Stream assistant output
      tags:
        - Assistant
      security:
        - {}
        - supabaseBearer: []
      responses:
        '200':
          description: SSE stream; inspect the error event because HTTP headers are already sent.
          content:
            text/event-stream:
              schema:
                type: string
              example: 'data: {"type":"meta","provider":"mock","mode":"demo","persisted":false}


                data: {"type":"delta","content":"AI not configured — demo response"}


                data: {"type":"done","provider":"mock","mode":"demo"}


                '
        '400':
          description: Invalid request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '429':
          description: Quota exceeded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          headers:
            Retry-After:
              schema:
                type: integer
              description: Seconds until retry.
        '503':
          description: Required credentials, migrations, or service are unavailable.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      description:
        'Optional bearer token. Guests always receive a deterministic labeled demo. Live OpenAI
        requires verified authentication and server service configuration. Sum of message content is limited
        to 48,000 characters. The last message must be user. Saved conversation history is loaded through RLS.
        SSE data JSON types: meta, delta, done, error. meta includes provider, mode, conversationId when saved,
        and persisted. Deltas use content. In-stream errors use code/message. Stop the request to cancel generation.
        Demo responses are not live AI. Anthropic/Gemini adapters are unimplemented. Quotas count request attempts;
        provider usage is recorded when returned.'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChatRequest'
  /api/ai/conversations:
    get:
      operationId: get_ai_conversations
      summary: List own conversations
      tags:
        - Assistant
      security:
        - supabaseBearer: []
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Conversation'
        '400':
          description: Invalid request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '429':
          description: Quota exceeded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          headers:
            Retry-After:
              schema:
                type: integer
              description: Seconds until retry.
        '503':
          description: Required credentials, migrations, or service are unavailable.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Missing, invalid, or expired Supabase bearer token.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Insufficient permission.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      parameters:
        - *id002
        - *id003
    post:
      operationId: post_ai_conversations
      summary: Create a conversation
      tags:
        - Assistant
      security:
        - supabaseBearer: []
      responses:
        '201':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Conversation'
        '400':
          description: Invalid request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '429':
          description: Quota exceeded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          headers:
            Retry-After:
              schema:
                type: integer
              description: Seconds until retry.
        '503':
          description: Required credentials, migrations, or service are unavailable.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Missing, invalid, or expired Supabase bearer token.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Insufficient permission.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                  minLength: 1
                  maxLength: 200
                  default: New conversation
              additionalProperties: false
  /api/ai/conversations/{id}:
    get:
      operationId: get_ai_conversations_id
      summary: Read a conversation
      tags:
        - Assistant
      security:
        - supabaseBearer: []
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Conversation'
        '400':
          description: Invalid request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '429':
          description: Quota exceeded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          headers:
            Retry-After:
              schema:
                type: integer
              description: Seconds until retry.
        '503':
          description: Required credentials, migrations, or service are unavailable.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Missing, invalid, or expired Supabase bearer token.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Insufficient permission.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      parameters:
        - name: id
          in: path
          required: true
          schema: *id001
    delete:
      operationId: delete_ai_conversations_id
      summary: Delete a conversation and messages
      tags:
        - Assistant
      security:
        - supabaseBearer: []
      responses:
        '204':
          description: Deleted or revoked; response body is empty.
        '400':
          description: Invalid request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '429':
          description: Quota exceeded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          headers:
            Retry-After:
              schema:
                type: integer
              description: Seconds until retry.
        '503':
          description: Required credentials, migrations, or service are unavailable.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Missing, invalid, or expired Supabase bearer token.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Insufficient permission.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      parameters:
        - name: id
          in: path
          required: true
          schema: *id001
  /api/ai/conversations/{id}/messages:
    get:
      operationId: get_ai_conversations_id_messages
      summary: Read conversation messages
      tags:
        - Assistant
      security:
        - supabaseBearer: []
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ChatMessage'
        '400':
          description: Invalid request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '429':
          description: Quota exceeded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          headers:
            Retry-After:
              schema:
                type: integer
              description: Seconds until retry.
        '503':
          description: Required credentials, migrations, or service are unavailable.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Missing, invalid, or expired Supabase bearer token.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Insufficient permission.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      parameters:
        - name: id
          in: path
          required: true
          schema: *id001
        - name: limit
          in: query
          schema:
            type: integer
            default: 200
            minimum: 1
            maximum: 500
        - *id003
...
components:
  securitySchemes:
    supabaseBearer:
      type: http
      scheme: bearer
      bearerFormat: Supabase access JWT
      description:
        A verified Supabase access token; API calls auth.getUser to verify it before creating a
        user-scoped database client.
  schemas:
    Error:
      type: object
      properties:
        error:
          type: object
          properties:
            code: *id004
            message: *id004
            requestId: *id004
            details: {}
          required:
            - code
            - message
          additionalProperties: false
      required:
        - error
      additionalProperties: false
...
    ChatMessage:
      type: object
      properties:
        id: *id001
        role:
          type: string
          enum:
            - user
            - assistant
        content:
          type: string
          minLength: 1
          maxLength: 16000
        provider: *id004
        mode:
          type: string
          enum:
            - live
            - demo
            - mock
        created_at:
          type: string
          format: date-time
      required:
        - role
        - content
      additionalProperties: false
    ChatRequest:
      type: object
      properties:
        messages:
          type: array
          minItems: 1
          maxItems: 30
          items:
            type: object
            properties:
              role:
                type: string
                enum:
                  - user
                  - assistant
              content:
                type: string
                minLength: 1
                maxLength: 16000
            required:
              - role
              - content
            additionalProperties: false
        context:
          type: object
          description:
            A CalculationResult with toolId and normalizedInput, or {toolId,input}. The API recomputes
            the attachment; arbitrary context is rejected.
          additionalProperties: true
        conversationId: *id001
      required:
        - messages
      additionalProperties: false
    Conversation:
      type: object
      properties:
        id: *id001
        title: *id004
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
      required:
        - id
        - title
        - created_at
      additionalProperties: false
...